+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.1 under development
 
 - New #142: Add `CurlTransport`.
+- New #144: Add `chatId` parameter to `SendGift` method.
 
 ## 0.7.0 February 10, 2025